framework 1.2.3

+++ wa-system/contact/waContactFields.class.php
@@ -282,10 +282,10 @@ class waContactFields
 
         // Remove field from order lists first
         if (isset(self::$companyFields[$id])) {
-            self::disableField($id, 'company');
+            self::disableField($id, 'company', true);
         }
         if (isset(self::$personFields[$id])) {
-            self::disableField($id, 'person');
+            self::disableField($id, 'person', true);
         }
 
         $file = wa()->getConfig()->getConfigPath('custom_fields.php', true, 'contacts');
@@ -457,9 +457,10 @@ class waContactFields
      * Remove given field from person or company order list.
      * @param $type string person|company
      * @param $id waContactField|int field ID or field instance.
+     * @param boolean $delete delete values from db or not
      * @throws waException
      */
-    public static function disableField($id, $type) {
+    public static function disableField($id, $type, $delete = false) {
         self::ensureStaticVars();
         if (is_object($id) && $id instanceof waContactField) {
             $id = $id->getId();
@@ -495,8 +496,10 @@ class waContactFields
          * @var waContactField $f
          */
 
-        // Remove data from DB
-        $f->getStorage()->deleteAll($id, $type);
+        if ($delete) {
+            // Remove data from DB
+            $f->getStorage()->deleteAll($id, $type);
+        }
 
         // Remove field from order file
         if (!is_readable($file)) {
